bench: move into the package so coverage and spawn both see it
Codex coverage blocker: the smoke test drove the script via
subprocess.run, which the coverage gate does not instrument (repo
precedent: review-checks-py drives logic unit-level for the same
reason) — the 349-line script measured 0.0%. The benchmark now lives at
ag2_sparrow.bench_claim_backends (importable: spawn workers re-import
it by name, and the smoke test drives main() IN-PROCESS under the
gate); scripts/bench-claim-backends.py stays as a thin CLI shim.
The smoke test's __main__ guard is load-bearing: spawn children
re-import the test module, and an unguarded module-level main() trips
_check_not_importing_main.
